Moved all browser initialisation into Driver object.  Introduced options for RemoteWebdriver, both chrome and firefox.
