--- src/main/g8/src/test/scala/uk/gov/hmrc/test/ui/driver/browsers/ChromeBrowser.scala
@@ -1,43 +1,0 @@
-/*
- * Copyright 2018 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package uk.gov.hmrc.test.ui.driver.browsers
-
-import org.openqa.selenium.WebDriver
-import org.openqa.selenium.chrome.{ChromeDriver, ChromeOptions}
-import org.openqa.selenium.remote.{CapabilityType, DesiredCapabilities}
-import uk.gov.hmrc.test.ui.driver.ProxySupport
-
-object ChromeBrowser extends ProxySupport {
-
-  def initialise(javascriptEnabled: Boolean, headlessMode: Boolean): WebDriver = {
-    val options = new ChromeOptions()
-    options.addArguments("test-type")
-    options.addArguments("--no-sandbox")
-    options.addArguments("start-maximized")
-    options.addArguments("disable-infobars")
-    val capabilities: DesiredCapabilities = DesiredCapabilities.chrome()
-    if (sys.props.get("qa.proxy").isDefined) {
-      capabilities.setCapability(CapabilityType.PROXY, initialiseProxy())
-    }
-    options.setCapability("javascript.enabled", javascriptEnabled)
-    options.merge(capabilities)
-    if (headlessMode) {
-      options.addArguments("headless")
-    }
-    new ChromeDriver(options)
-  }
-
-}
