--- src/main/g8/src/test/scala/uk/gov/hmrc/test/ui/driver/browsers/FirefoxBrowser.scala
@@ -1,41 +1,0 @@
-/*
- * Copyright 2018 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package uk.gov.hmrc.test.ui.driver.browsers
-
-import org.openqa.selenium.WebDriver
-import org.openqa.selenium.firefox.{FirefoxDriver, FirefoxOptions, FirefoxProfile}
-import org.openqa.selenium.remote.DesiredCapabilities
-
-object FirefoxBrowser {
-
-  def initialise(javascriptEnabled: Boolean): WebDriver = {
-    System.setProperty(FirefoxDriver.SystemProperty.DRIVER_USE_MARIONETTE, "true")
-    System.setProperty(FirefoxDriver.SystemProperty.BROWSER_LOGFILE, "/dev/null")
-    val profile = new FirefoxProfile()
-    profile.setAcceptUntrustedCertificates(true)
-    profile.setPreference("javascript.enabled", javascriptEnabled)
-
-    val options = new FirefoxOptions()
-    val capabilities = DesiredCapabilities.firefox()
-
-    options.merge(capabilities)
-    options.setProfile(profile)
-    options.setAcceptInsecureCerts(true)
-
-    new FirefoxDriver(options)
-  }
-
-}
